Refactor header: Extract header component from BlogApp and PostDetail, implement formatDate utility function for date formatting, and update Contact component text.
