+++ src/utilities/formatDate.js
@@ -1,0 +1,10 @@
+// 日付フォーマット関数
+const formatDate = (dateString) => {
+  const date = new Date(dateString);
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  const day = date.getDate();
+  return `${year}/${month}/${day}`;
+};
+
+export default formatDate;
